requirements.txt
airtable-python-wrapper==0.15.2
appnope==0.1.2
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
decorator==5.0.6
Flask==1.1.2
idna==2.10
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jupyter-client==6.1.12
jupyter-core==4.7.1
MarkupSafe==1.1.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.8.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
pyzmq==22.0.3
requests==2.25.1
six==1.15.0
tornado==6.1
traitlets==5.0.5
twilio==6.56.0
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
